requirements.txt
fastapi>=0.111,<1.0
uvicorn[standard]>=0.30,<1.0
httpx>=0.27,<1.0
pydantic>=2.0,<3.0
sse-starlette>=2.0,<3.0
Pillow>=10.0,<12.0
python-multipart>=0.0.9,<1.0
pyyaml>=6.0,<7.0
pytest>=8.0,<9.0
pytest-asyncio>=0.23,<1.0
ruff>=0.4,<1.0
mypy>=1.10,<2.0
aiofiles>=23.0,<25.0
respx